+++ app/blog/page.tsx
@@ -4,8 +4,8 @@ import { useState, useEffect } from "react";
 import Link from "next/link";
 import { motion } from "motion/react";
 import { ArrowLeft } from "lucide-react";
-import { BlogCard } from "@/components/BlogCard";
-import { blogPosts } from "@/data/blog-posts";
+import { BlogCard } from "../../components/BlogCard";
+import { blogPosts } from "../../src/data/blog-posts";
 
 const categories = [
   "All",
