+++ src/components/Companies.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { InfiniteMovingCards } from "./ui/infinite-moving-cards";
+import { InfiniteMovingCards } from "@/components/ui/infinite-moving-cards";
 
 export default function Companies() {
   const logos = [
