+++ src/components/ui/direction-aware-hover.tsx
@@ -3,7 +3,7 @@
 import { useRef, useState } from "react";
 
 import { AnimatePresence, motion } from "motion/react";
-import { cn } from "../../lib/utils";
+import { cn } from "@/lib/utils";
 
 export const DirectionAwareHover = ({
   imageUrl,
