+++ src/components/ui/text-generate-effect.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { useEffect, useState } from "react";
 import { motion, stagger, useAnimate } from "motion/react";
-import { cn } from "../../lib/utils";
+import { cn } from "@/lib/utils";
 
 export const TextGenerateEffect = ({
   words,
